Iris#2 CRUD user (#5)
🚀 This MR closes iris #2 CRUD user

📌 Description
Development of the registration and login using the API.

👾 Tech details
For now everything is working only locally, the backend is configured for localhost:3000 and connected to the frontend on localhost:3001.

📝 Steps to reproduce
For everything to work, you must:

Run npm run dev:start in a terminal to initialize the backend.
Once started, you can create, log-in, update, delete any user.